prototype 2

+++ 0-bubble_sort.c
@@ -12,7 +12,7 @@
 void bubble_sort(int *array, size_t size)
 {
 	size_t i, j;
-	size_t temp;
+	int temp;
 
 	if (size < 1 || !array)
 		return;
@@ -22,7 +22,7 @@ void bubble_sort(int *array, size_t size)
 		for (j = 0; j < size - 1; j++)
 		{
 			if (array[j] > array[j + 1])
-			{
+			{	/*swap values of elements*/
 				temp = array[j];
 				array[j] = array[j + 1];
 				array[j + 1] = temp;
